requirements.txt
aniso8601==9.0.1
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pytz==2021.3
six==1.16.0
Werkzeug==2.0.2

requests~=2.27.1
Pillow~=9.0.0